Add variable for path to sudoers.d
On some systems (e.g. BSDs) where sudo is not part of the base system, the
sudoers.d directory is at /usr/local/etc/sudoers.d rather than /etc/sudoers.d.
This commit adds a variable to allow users of the role to specify an alternate
path for the sudoers.d directory.