Added additional Default configuration options
This is according to the sudoers manpage.
Default_Type ::= 'Defaults' |
           'Defaults' '@' Host_List |
           'Defaults' ':' User_List |
           'Defaults' '!' Cmnd_List |
           'Defaults' '>' Runas_List
Host_List is not available as I see no need for this configuration option if we make use of ansible.
The default if is a user list if name is defined, otherwise it is changed according to type which can be user, cmnd or runas.